Se agrega validaciones y se pasa el tipo de listas a Double

+++ src/calculator/Main.java
@@ -6,8 +6,8 @@ import java.util.stream.Collectors;
 public class Main {
     public static void main(String[] args) {
         Operations operations = new Operations();
-            List<Integer> listOne = List.of(1,2,3,10,25,6);
-            List<Integer> listTwo = List.of(1,5,3,4,4,6);
+            List<Double> listOne = List.of(1.0,2.0,3.0,10.0,25.0,6.0);
+            List<Double> listTwo = List.of(1.0,5.0,3.0,4.0,4.0,6.0);
         System.out.println("LISTAS");
         System.out.println(listOne.stream().collect(Collectors.toList()));
         System.out.println(listTwo.stream().collect(Collectors.toList()));
